refactor: addresses non-sibling use of "/definition" for all non-"exports" files
