+++ src/library/UniformResourceIdentifier/definition.ts
@@ -1,5 +1,5 @@
 import Attempt from '@/library/Attempt';
-import type NonTrivialString from '@/library/NonTrivialString/definition';
+import type NonTrivialString from '@/library/NonTrivialString/exports';
 
 import {
   concatenated,
